+++ src/test/java/org/idnt/udemy/mockitoapp/example/service/impl/test/ExamServiceImplTest.java
@@ -5,6 +5,7 @@ import org.idnt.udemy.mockitoapp.example.repository.ExamRepository;
 import org.idnt.udemy.mockitoapp.example.repository.QuestionRepository;
 import org.idnt.udemy.mockitoapp.example.repository.impl.ExamRepositoryImpl;
 import org.idnt.udemy.mockitoapp.example.repository.impl.QuestionRepositoryImpl;
+import org.idnt.udemy.mockitoapp.example.service.ExamService;
 import org.idnt.udemy.mockitoapp.example.service.impl.ExamServiceImpl;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -292,6 +293,7 @@ class ExamServiceImplTest {
     }
 
     @Test
+    @DisplayName("Test \"doCallRealMethod()\"")
     void testDoCallRealMethod() {
         //Given
         when(this.examRepository.findAll()).thenReturn(this.dataListExam);
@@ -311,7 +313,34 @@ class ExamServiceImplTest {
         assertEquals("Matemáticas", exam.getName(), () -> "The name of the exam doesn't match with the expected name.");
         assertNotNull(exam.getQuestions(), () -> "The questions of resulting exam can't be null.");
         assertEquals(0, exam.getQuestions().size(), () ->
-                String.format("The size of the questions of resulting exam must be %s", 0));
+                String.format("The size of the questions of resu lting exam must be %s", 0));
         verify(this.questionRepository).findQuestionByExamId(anyLong());
+    }
+
+    @Test
+    @DisplayName("Test \"spy()\"")
+    void testSpy() {
+        //Given
+        ExamRepository examRepository = spy(ExamRepositoryImpl.class);
+        QuestionRepository questionRepository = spy(QuestionRepositoryImpl.class);
+        ExamService examService = new ExamServiceImpl(examRepository, questionRepository);
+
+        /**
+         * In the "when" method is being called because it is a real method.
+         * Better to use the "doReturn" method
+         */
+        //when(questionRepository.findQuestionByExamId(anyLong())).thenReturn(Collections.emptyList());
+        doReturn(Collections.emptyList()).when(questionRepository).findQuestionByExamId(anyLong());
+
+        //When
+        Optional<Exam> examWithQuestionsOptional = examService.findExamByNameWithQuestions("Matemáticas");
+
+        //Then
+        assertTrue(examWithQuestionsOptional.isPresent(), () -> "The resulting exam can't be void.");
+        Exam exam = examWithQuestionsOptional.get();
+        assertEquals("Matemáticas", exam.getName(), () -> "The name of the exam doesn't match with the expected name.");
+        assertNotNull(exam.getQuestions(), () -> "The questions of resulting exam can't be null.");
+        assertEquals(0, exam.getQuestions().size(), () ->
+                String.format("The size of the questions of resu lting exam must be %s", 0));
     }
 }